refactor(state): moving state persistence and actions into a dedicated module
