+++ src/fulgur/state/mod.rs
@@ -1,0 +1,10 @@
+mod operations;
+mod persistence;
+mod writer;
+
+pub use operations::{TabRestoreDecision, determine_tab_restore_strategy};
+pub use persistence::{
+    SerializedRemoteSpec, SerializedWindowBounds, TabState, WindowState, WindowsState,
+    get_file_modified_time, is_file_newer,
+};
+pub use writer::StateWriter;
